refactor(host): share child worker supervision
Generated-By: pi 0.84.3 (https://pi.dev)
